Revert "Réorganisation de quelques fonctions métier + calcul et affichage de l'équilibre budgétaire (#73)"
This reverts commit b11735195c7d5361698d2c21fb503d33622319cd.
