+++ src/public/js/components/TopLevel.js
@@ -1,11 +1,10 @@
 import React from 'react';
-import {format} from 'currency-formatter';
 
 
 export default function({expenditures, revenue}){
 
     return true ? React.createElement('ul', {}, 
-        React.createElement('li', {}, 'Dépenses : ', format(expenditures, { code: 'EUR' })),
-        React.createElement('li', {}, 'Recettes : ', format(revenue, { code: 'EUR' }))
+        React.createElement('li', {}, 'Dépenses : ', expenditures),
+        React.createElement('li', {}, 'Recettes : ', revenue)
     ) : React.createElement('div', {});
 }
