+++ src/public/js/mapStateToProps.js
@@ -1,7 +1,6 @@
-import budgetBalance from '../../shared/js/finance/budgetBalance';
-
 export default function(state){
-    const { m52Instruction } = state;
-
-    return m52Instruction ? budgetBalance(m52Instruction) : {};
+    return {
+        expenditures: 1000,
+        revenue: 1002
+    };
 }
